final_update
