--- SummerProject/article/forms.py
@@ -1,9 +1,0 @@
-from django import forms
-from .models import Comment
-
-
-class CommentForm(forms.ModelForm):
-
-    class Meta:
-        model = Comment
-        fields = ['content']
