+++ SummerProject/articles/views.py
@@ -1,9 +1,13 @@
 from django.utils import timezone
-from django.http import JsonResponse, HttpResponse
+from django.http import JsonResponse, HttpResponse, HttpResponseBadRequest,
 from django.views import View
 from articles.models import Article
 from django.forms.models import model_to_dict
 from django.shortcuts import get_object_or_404
+from django.db import transaction
+import json
+from django.db.models import Count
+from .forms import CommentForm
 
 
 class ArticleDetailView(View):
@@ -14,3 +18,36 @@ class ArticleDetailView(View):
     def get(self, request, pk):
         article = get_object_or_404(Article, pk=pk)
         return JsonResponse(model_to_dict(article))
+
+""" View for comment backend"""
+@transaction.atomic
+def comments_view(request, article_id):
+    article = get_object_or_404(Article, id=article_id)
+    if request.method == "GET":
+        data = list(article.comments.all().values())
+        return JsonResponse(data, safe=False)
+    elif request.method == 'POST':
+        data = request.body.decode('utf8')
+        data = json.loads(data)
+        form = CommentForm(data)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.user = request.user
+            comment.save()
+            article.comments.add(comment)
+            response = HttpResponse(status=201)
+            response['comment_id'] = comment.id
+            response['article_id'] = article_id
+            return response
+    else:
+        return HttpResponseBadRequest
+
+
+""" View for most commented articles"""
+
+
+@transaction.atomic
+def most_commented(request):
+    if request.method == "GET":
+        data = list(Article.objects.values().annotate(art_comments=Count('comments')).order_by('-art_comments')[:3])
+        return JsonResponse(data, safe=False)
